requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
bleach==3.2.1
bs4==0.0.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
cycler==0.10.0
decorator==4.4.2
idna==2.10
kiwisolver==1.3.1
networkx==2.5
numpy==1.19.4
packaging==20.7
peewee==3.14.0
Pillow==8.0.1
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-telegram-bot==13.1
pytz==2020.4
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.2
webencodings==0.5.1
